move cgmlp and fastformer to a separate directory
