+++ espnet2/asr/encoder/branchformer_encoder.py
@@ -1,13 +1,14 @@
-# Copyright 2022 Yifan Peng
+# Copyright 2022 Yifan Peng (Carnegie Mellon University)
 #  Apache 2.0  (http://www.apache.org/licenses/LICENSE-2.0)
 
 """Branchformer encoder definition.
 
 Reference:
-    Yifan Peng, Siddharth Dalmia, Ian Lane, and Shinji Watanabe,
-    "Branchformer: Parallel MLP-Attention Architectures to
-    Capture Local and Global Context for Speech Recognition
-    and Understanding," in ICML 2022.
+    Yifan Peng, Siddharth Dalmia, Ian Lane, and Shinji Watanabe, 
+    “Branchformer: Parallel MLP-Attention Architectures to Capture 
+    Local and Global Context for Speech Recognition and Understanding,” 
+    in Proceedings of ICML, 2022.
+
 """
 
 import logging
@@ -18,6 +19,8 @@ import torch
 from typeguard import check_argument_types
 
 from espnet2.asr.encoder.abs_encoder import AbsEncoder
+from espnet2.asr.layers.cgmlp import ConvolutionalGatingMLP
+from espnet2.asr.layers.fastformer import FastSelfAttention
 from espnet.nets.pytorch_backend.nets_utils import get_activation, make_pad_mask
 from espnet.nets.pytorch_backend.transformer.attention import (  # noqa: H301
     LegacyRelPositionMultiHeadedAttention,
@@ -40,278 +43,6 @@ from espnet.nets.pytorch_backend.transformer.subsampling import (
     TooShortUttError,
     check_short_utt,
 )
-
-
-class ConvolutionalSpatialGatingUnit(torch.nn.Module):
-    """Convolutional Spatial Gating Unit (CSGU).
-
-    References:
-        - https://openreview.net/forum?id=RA-zVvZLYIy
-        - https://arxiv.org/abs/2105.08050
-    """
-
-    def __init__(
-        self,
-        size: int,
-        kernel_size: int,
-        dropout_rate: float,
-        use_linear_after_conv: bool,
-        gate_activation: str,
-    ):
-        super().__init__()
-
-        n_channels = size // 2  # split input channels
-        self.norm = LayerNorm(n_channels)
-        self.conv = torch.nn.Conv1d(
-            n_channels,
-            n_channels,
-            kernel_size,
-            1,
-            (kernel_size - 1) // 2,
-            groups=n_channels,
-        )
-        if use_linear_after_conv:
-            self.linear = torch.nn.Linear(n_channels, n_channels)
-        else:
-            self.linear = None
-
-        if gate_activation == "identity":
-            self.act = torch.nn.Identity()
-        else:
-            self.act = get_activation(gate_activation)
-
-        self.dropout = torch.nn.Dropout(dropout_rate)
-
-    def espnet_initialization_fn(self):
-        torch.nn.init.normal_(self.conv.weight, std=1e-6)
-        torch.nn.init.ones_(self.conv.bias)
-        if self.linear is not None:
-            torch.nn.init.normal_(self.linear.weight, std=1e-6)
-            torch.nn.init.ones_(self.linear.bias)
-
-    def forward(self, x, gate_add=None):
-        """Forward method
-
-        Args:
-            x (torch.Tensor): (N, T, D)
-            gate_add (torch.Tensor): (N, T, D/2)
-
-        Returns:
-            out (torch.Tensor): (N, T, D/2)
-        """
-
-        x_r, x_g = x.chunk(2, dim=-1)
-
-        x_g = self.norm(x_g)  # (N, T, D/2)
-        x_g = self.conv(x_g.transpose(1, 2)).transpose(1, 2)  # (N, T, D/2)
-        if self.linear is not None:
-            x_g = self.linear(x_g)
-
-        if gate_add is not None:
-            x_g = x_g + gate_add
-
-        x_g = self.act(x_g)
-        out = x_r * x_g  # (N, T, D/2)
-        out = self.dropout(out)
-        return out
-
-
-class ConvolutionalGatingMLP(torch.nn.Module):
-    """Convolutional Gating MLP (cgMLP).
-
-    Note: LayerNorm and skip connection should be applied outside this module.
-
-    References:
-        - https://openreview.net/forum?id=RA-zVvZLYIy
-        - https://arxiv.org/abs/2105.08050
-    """
-
-    def __init__(
-        self,
-        size: int,
-        linear_units: int,
-        kernel_size: int,
-        dropout_rate: float,
-        use_linear_after_conv: bool,
-        gate_activation: str,
-    ):
-        super().__init__()
-
-        self.channel_proj1 = torch.nn.Sequential(
-            torch.nn.Linear(size, linear_units), torch.nn.GELU()
-        )
-        self.csgu = ConvolutionalSpatialGatingUnit(
-            size=linear_units,
-            kernel_size=kernel_size,
-            dropout_rate=dropout_rate,
-            use_linear_after_conv=use_linear_after_conv,
-            gate_activation=gate_activation,
-        )
-        self.channel_proj2 = torch.nn.Linear(linear_units // 2, size)
-
-    def forward(self, x, mask):
-        if isinstance(x, tuple):
-            xs_pad, pos_emb = x
-        else:
-            xs_pad, pos_emb = x, None
-
-        xs_pad = self.channel_proj1(xs_pad)  # size -> linear_units
-        xs_pad = self.csgu(xs_pad)  # linear_units -> linear_units/2
-        xs_pad = self.channel_proj2(xs_pad)  # linear_units/2 -> size
-
-        if pos_emb is not None:
-            out = (xs_pad, pos_emb)
-        else:
-            out = xs_pad
-        return out
-
-
-class FastSelfAttention(torch.nn.Module):
-    """Fast self-attention used in Fastformer.
-
-    Reference:
-        Wu et al., "Fastformer: Additive Attention Can Be All You Need"
-        https://arxiv.org/abs/2108.09084
-        Code is based on: https://github.com/wuch15/Fastformer
-    """
-
-    def __init__(
-        self,
-        size,
-        attention_heads,
-        dropout_rate,
-    ):
-        super().__init__()
-        if size % attention_heads != 0:
-            raise ValueError(
-                f"Hidden size ({size}) is not an integer multiple "
-                f"of attention heads ({attention_heads})"
-            )
-        self.attention_head_size = size // attention_heads
-        self.num_attention_heads = attention_heads
-
-        self.query = torch.nn.Linear(size, size)
-        self.query_att = torch.nn.Linear(size, attention_heads)
-        self.key = torch.nn.Linear(size, size)
-        self.key_att = torch.nn.Linear(size, attention_heads)
-        self.transform = torch.nn.Linear(size, size)
-        self.dropout = torch.nn.Dropout(dropout_rate)
-
-    def espnet_initialization_fn(self):
-        self.apply(self.init_weights)
-
-    def init_weights(self, module):
-        if isinstance(module, torch.nn.Linear):
-            module.weight.data.normal_(mean=0.0, std=0.02)
-        if isinstance(module, torch.nn.Linear) and module.bias is not None:
-            module.bias.data.zero_()
-
-    def transpose_for_scores(self, x):
-        """Reshape and transpose to compute scores.
-
-        Args:
-            x: (batch, time, size = n_heads * attn_dim)
-
-        Returns:
-            (batch, n_heads, time, attn_dim)
-        """
-
-        new_x_shape = x.shape[:-1] + (
-            self.num_attention_heads,
-            self.attention_head_size,
-        )
-        return x.reshape(*new_x_shape).transpose(1, 2)
-
-    def forward(self, xs_pad, mask):
-        """Forward method.
-
-        Args:
-            xs_pad: (batch, time, size = n_heads * attn_dim)
-            mask: (batch, 1, time), nonpadding is 1, padding is 0
-
-        Returns:
-            torch.Tensor: (batch, time, size)
-        """
-
-        batch_size, seq_len, _ = xs_pad.shape
-        mixed_query_layer = self.query(xs_pad)  # (batch, time, size)
-        mixed_key_layer = self.key(xs_pad)  # (batch, time, size)
-
-        if mask is not None:
-            mask = mask.eq(0)  # padding is 1, nonpadding is 0
-
-        # (batch, n_heads, time)
-        query_for_score = (
-            self.query_att(mixed_query_layer).transpose(1, 2)
-            / self.attention_head_size**0.5
-        )
-        if mask is not None:
-            min_value = float(
-                numpy.finfo(
-                    torch.tensor(0, dtype=query_for_score.dtype).numpy().dtype
-                ).min
-            )
-            query_for_score = query_for_score.masked_fill(mask, min_value)
-            query_weight = torch.softmax(query_for_score, dim=-1).masked_fill(mask, 0.0)
-        else:
-            query_weight = torch.softmax(query_for_score, dim=-1)
-
-        query_weight = query_weight.unsqueeze(2)  # (batch, n_heads, 1, time)
-        query_layer = self.transpose_for_scores(
-            mixed_query_layer
-        )  # (batch, n_heads, time, attn_dim)
-
-        pooled_query = (
-            torch.matmul(query_weight, query_layer)
-            .transpose(1, 2)
-            .reshape(-1, 1, self.num_attention_heads * self.attention_head_size)
-        )  # (batch, 1, size = n_heads * attn_dim)
-        pooled_query = self.dropout(pooled_query)
-        pooled_query_repeat = pooled_query.repeat(1, seq_len, 1)  # (batch, time, size)
-
-        mixed_query_key_layer = (
-            mixed_key_layer * pooled_query_repeat
-        )  # (batch, time, size)
-
-        # (batch, n_heads, time)
-        query_key_score = (
-            self.key_att(mixed_query_key_layer) / self.attention_head_size**0.5
-        ).transpose(1, 2)
-        if mask is not None:
-            min_value = float(
-                numpy.finfo(
-                    torch.tensor(0, dtype=query_key_score.dtype).numpy().dtype
-                ).min
-            )
-            query_key_score = query_key_score.masked_fill(mask, min_value)
-            query_key_weight = torch.softmax(query_key_score, dim=-1).masked_fill(
-                mask, 0.0
-            )
-        else:
-            query_key_weight = torch.softmax(query_key_score, dim=-1)
-
-        query_key_weight = query_key_weight.unsqueeze(2)  # (batch, n_heads, 1, time)
-        key_layer = self.transpose_for_scores(
-            mixed_query_key_layer
-        )  # (batch, n_heads, time, attn_dim)
-        pooled_key = torch.matmul(
-            query_key_weight, key_layer
-        )  # (batch, n_heads, 1, attn_dim)
-        pooled_key = self.dropout(pooled_key)
-
-        # NOTE: value = query, due to param sharing
-        weighted_value = (pooled_key * query_layer).transpose(
-            1, 2
-        )  # (batch, time, n_heads, attn_dim)
-        weighted_value = weighted_value.reshape(
-            weighted_value.shape[:-2]
-            + (self.num_attention_heads * self.attention_head_size,)
-        )  # (batch, time, size)
-        weighted_value = (
-            self.dropout(self.transform(weighted_value)) + mixed_query_layer
-        )
-
-        return weighted_value
 
 
 class BranchformerEncoderLayer(torch.nn.Module):
